changed parameters folder name
